+++ app/Models/User.php
@@ -2,13 +2,9 @@
 
 namespace App\Models;
 
-// use Illuminate\Contracts\Auth\MustVerifyEmail;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
-use Illuminate\Support\Facades\Cache;
 use Laravel\Sanctum\HasApiTokens;
 
 /**
@@ -45,6 +41,7 @@ class User extends Authenticatable
     use HasApiTokens;
     use HasFactory;
     use Notifiable;
+    use HasPermissions;
 
     protected $fillable = [
         'name',
@@ -61,37 +58,4 @@ class User extends Authenticatable
         'email_verified_at' => 'datetime',
         'password'          => 'hashed',
     ];
-
-    public function permissions(): BelongsToMany
-    {
-        return $this->belongsToMany(Permission::class);
-    }
-
-    public function givePermissionTo(string $key): void
-    {
-        $this->permissions()->firstOrCreate(compact('key'));
-
-        Cache::forget($this->getPermissionCacheKey());
-
-        Cache::rememberForever(
-            $this->getPermissionCacheKey(),
-            fn () => $this->permissions
-        );
-
-    }
-
-    public function hasPermissionTo(string $key): bool
-    {
-        /** @var Collection $permissions  */
-        $permissions = Cache::get($this->getPermissionCacheKey(), $this->permissions);
-
-        return $permissions
-            ->where('key', '=', $key)
-            ->isNotEmpty();
-    }
-
-    public function getPermissionCacheKey(): string
-    {
-        return "user::{$this->id}::permissions";
-    }
 }
